+++ plinth/modules/first_boot/forms.py
@@ -33,7 +33,7 @@ from plinth import actions
 from plinth import cfg
 from plinth.errors import ActionError, DomainRegistrationError
 from plinth.modules.pagekite.utils import PREDEFINED_SERVICES, run
-from plinth.modules.security.views import set_restricted_access
+from plinth.modules.security import set_restricted_access
 from plinth.modules.users.forms import GROUP_CHOICES
 from plinth.utils import format_lazy
 
@@ -78,7 +78,15 @@ class State1Form(auth.forms.UserCreationForm):
                             self.cleaned_data['password1'])
 
             # Restrict console login to users in admin or sudo group
-            set_restricted_access(True)
+            try:
+                set_restricted_access(True)
+                message = _('Console login access restricted to users in '
+                            '"admin" group. This can be configured in '
+                            'security settings.')
+                messages.success(self.request, message)
+            except Exception:
+                messages.error(self.request,
+                               _('Failed to restrict console access.'))
 
         return user
