+++ plinth/modules/security/templates/security.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "simple_service.html" %}
 {% comment %}
 #
 # This file is part of Plinth.
@@ -21,7 +21,7 @@
 {% load bootstrap %}
 {% load i18n %}
 
-{% block content %}
+{% block configuration %}
 
   <form class="form" method="post">
     {% csrf_token %}
